# Simplifications and performance improvements for estimators