--- a/src/estimator.jl
+++ b/src/estimator.jl
@@ -2,26 +2,45 @@
 #  * The input is nonempty
 #  * Time 0 is not included
 
-function _estimator(::Type{S}, tte::AbstractVector{T}, status::BitVector) where {S, T}
-    nobs = length(tte)
+function _n_unique_times(ets)
+    # We know that `ets` is nonempty and sorted ascending by the `time` fields of the
+    # elements, so we can count the unique values in `O(length(ets))` time with no
+    # allocations by only comparing to the last observed unique time
+    t_prev = first(ets).time
+    n = 1
+    for et in Iterators.drop(ets, 1)
+        t = et.time
+        if t != t_prev
+            n += 1
+            t_prev = t
+        end
+    end
+    return n
+end
+
+function _estimator(::Type{S}, ets::AbstractVector{EventTime{T}}) where {S,T}
+    outlen = _n_unique_times(ets)
+
+    nobs = length(ets)
     dᵢ = 0                   # Number of observed events at time t
     cᵢ = 0                   # Number of censored events at time t
     nᵢ = nobs                # Number remaining at risk at time t
     es = estimator_start(S)  # Estimator starting point
     gw = stderr_start(S)     # Standard Error starting point
 
-    times = T[]              # The set of unique event times
-    nevents = Int[]          # Total observed events at each time
-    ncensor = Int[]          # Total censored events at each time
-    natrisk = Int[]          # Number at risk at each time
-    estimator = Float64[]    # Estimates
-    stderr = Float64[]       # Pointwise standard errors
+    times = Vector{T}(undef, outlen)            # The set of unique event times
+    nevents = Vector{Int}(undef, outlen)        # Total observed events at each time
+    ncensor = Vector{Int}(undef, outlen)        # Total censored events at each time
+    natrisk = Vector{Int}(undef, outlen)        # Number at risk at each time
+    estimator = Vector{Float64}(undef, outlen)  # Estimates
+    stderr = Vector{Float64}(undef, outlen)     # Pointwise standard errors
 
     t_prev = zero(T)
+    outind = 1
 
-    @inbounds for i = 1:nobs
-        t = tte[i]
-        s = status[i]
+    @inbounds for obsind in eachindex(ets)
+        t = ets[obsind].time
+        s = ets[obsind].status
         # Aggregate over tied times
         if t == t_prev
             dᵢ += s
@@ -30,12 +49,13 @@ function _estimator(::Type{S}, tte::AbstractVector{T}, status::BitVector) where 
         elseif !iszero(t_prev)
             es = estimator_update(S, es, dᵢ, nᵢ)
             gw = stderr_update(S, gw, dᵢ, nᵢ)
-            push!(times, t_prev)
-            push!(nevents, dᵢ)
-            push!(ncensor, cᵢ)
-            push!(natrisk, nᵢ)
-            push!(estimator, es)
-            push!(stderr, sqrt(gw))
+            times[outind] = t_prev
+            nevents[outind] = dᵢ
+            ncensor[outind] = cᵢ
+            natrisk[outind] = nᵢ
+            estimator[outind] = es
+            stderr[outind] = sqrt(gw)
+            outind += 1
         end
         nᵢ -= dᵢ + cᵢ
         dᵢ = s
@@ -45,12 +65,12 @@ function _estimator(::Type{S}, tte::AbstractVector{T}, status::BitVector) where 
 
     # We need to do this one more time to capture the last time
     # since everything in the loop is lagged
-    push!(times, t_prev)
-    push!(nevents, dᵢ)
-    push!(ncensor, cᵢ)
-    push!(natrisk, nᵢ)
-    push!(estimator, es)
-    push!(stderr, sqrt(gw))
+    times[outind] = t_prev
+    nevents[outind] = dᵢ
+    ncensor[outind] = cᵢ
+    natrisk[outind] = nᵢ
+    estimator[outind] = es
+    stderr[outind] = sqrt(gw)
 
     return S{T}(times, nevents, ncensor, natrisk, estimator, stderr)
 end
@@ -58,26 +78,16 @@ end
 function StatsBase.fit(::Type{S},
                        times::AbstractVector{T},
                        status::AbstractVector{<:Integer}) where {S<:NonparametricEstimator,T}
-    nobs = length(times)
-    if length(status) != nobs
-        throw(DimensionMismatch("there must be as many event statuses as times"))
+    ntimes = length(times)
+    nstatus = length(status)
+    if ntimes != nstatus
+        throw(DimensionMismatch("number of event statuses does not match number of " *
+                                "event times; got $nstatus and $ntimes, respectively"))
     end
-    if nobs == 0
-        throw(ArgumentError("the sample must be nonempty"))
-    end
-    p = sortperm(times)
-    t = times[p]
-    s = BitVector(status[p])
-    return _estimator(S, t, s)
+    return fit(S, map(EventTime{T}, times, status))
 end
 
 function StatsBase.fit(::Type{S}, ets::AbstractVector{<:EventTime}) where S<:NonparametricEstimator
-    length(ets) > 0 || throw(ArgumentError("the sample must be nonempty"))
-    x = sort(ets)
-    # TODO: Refactor, since iterating over the EventTime objects directly in
-    # the _km loop may actually be easier/more efficient than working with
-    # the times and statuses as separate vectors. Plus it might be nice to
-    # make this method the One True Method™ so that folks are encouraged to
-    # use EventTimes instead of raw values.
-    return fit(S, map(t->t.time, x), BitVector(map(t->t.status, x)))
+    isempty(ets) && throw(ArgumentError("can't compute $(nameof(S)) from 0 observations"))
+    return _estimator(S, issorted(ets) ? ets : sort(ets))
 end