--- a/src/kaplanmeier.jl
+++ b/src/kaplanmeier.jl
@@ -49,7 +49,15 @@ end
     fit(KaplanMeier, times, status) -> KaplanMeier
 
 Given a vector of times to events and a corresponding vector of indicators that
-dictate whether each time is an observed event or is right censored, compute the
+denote whether each time is an observed event or is right censored, compute the
 Kaplan-Meier estimate of the survivor function.
 """
 StatsBase.fit(::Type{KaplanMeier}, times, status)
+
+"""
+    fit(KaplanMeier, ets) -> KaplanMeier
+
+Compute the Kaplan-Meier estimate of the survivor function from a vector of
+[`EventTime`](@ref) values.
+"""
+StatsBase.fit(::Type{KaplanMeier}, ets)